requirements.txt
# Core dependencies
requests>=2.28.0
anthropic>=0.5.0
azure-identity>=1.12.0
azure-mgmt-resource>=21.1.0
azure-core>=1.26.0


# Web interface
flask>=2.2.0
flask-session>=0.4.0

# Utilities
python-dotenv>=0.19.0

# CLI interface
rich>=12.0.0
prompt-toolkit>=3.0.30

# Logging and utilities
pydantic>=1.10.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0

# Terraform execution
PyYAML>=6.0.0
jinja2>=3.0.0

#docker
gunicorn>=20.1.0